[IMP] project_sale_subscription: allow clicking on subscription section
Before this commit, the user can only see in the project profitability
made with the subscriptions related to the project but he cannot see
which subscriptions are related to this project.

This commit allows the user to click on the subscription section to see
the subscriptions related to this section and thus to the project.

task-2710808