[FIX] *: adapt some html fields in views
Some html field are not in their ideal style.
As the style for html fields is now dependent of
where you are in the xml view ( in group or not),
we have to adapt some views and flags some fields
to ensure they have the correct look.

This change is only be a visual enhancement
and should not prevent the function of said html fields
event if the views are not updated.

task-2637488

closes odoo/enterprise#23911

X-original-commit: 7f6011bf80e070a369bb3181c26c6a83cd3d030c
Related: odoo/odoo#83792
Signed-off-by: Antoine Guenet <age@odoo.com>
Signed-off-by: Geelen Sébastien (sge) <sge@odoo.com>