[FIX] *_mrp_workorder_*: tablet view refactoring
This commit redo the "tablet view" of workorder with a new client
action. The main component is now an action calling an Owl component
made of two small form views (one for the workorder, one for the current
quality check) and some widgets. The main feature is the ability to see
every quality checks to do (the `Step` component) directly on the tablet
view.

The current quality check has its own view so we get rid of a bunch of
related fields on mrp_workorder.

Task: 2427466
Part-of: odoo/enterprise#23032