[FIX] sign: SMS dialog receiving wrong parameters
In the SMSSignerDialog the signature process is made from a widget that
after filled and sent, finishes the signing process. However, when refactoring
this widget, the signature parameters was being passed wrongly. This blocks the
signing of any request that requires SMS validation. This commit fixes this issue
by passing the correct parameters to the init of the dialog.

closes odoo/enterprise#23751

Signed-off-by: Arnaud Joset <arj@odoo.com>