[FIX] sale_subscription: translation of empty description...
... due to adding values in demo data.

Adding the module sale_subscription will add the DATA yearly_subscription
sale_subscription_template) in the xml file of the DATA, we can see that
the dscription field is not populated.

However, in the xml file of the DEMO data, the yearly_subscription is
populated.

If you create a db without the demo data and add a language, the translation
of the yearly_subscription template for the description will be populated
however we did not have a value.

After discussing with our Translation expert, we should not modify the same
data in DATA and DEMO. This would lead to translation issue like the current one.

To solve the issue, let's simply remove the translation of the description
in order to avoid seeing terms being translated upon language adding.

opw-2692782 (where customer was upset thinking somebody was adding
translation on his db)

closes odoo/enterprise#23853

X-original-commit: 75bdfbcad1bc333b3d1bd8061ee4f0b17b631b40
Signed-off-by: Vergote Baptiste (bve) <bve@odoo.com>