[FIX] hr_contract: Exclude gaps on first contract date computation
Purpose
=======

This field is used for several things, like the simple holiday pay recovery.

Normally, we expect the HR officers to create a new employee, instead of
unarchiving the old one, but this could happen.

If someone leaves the company, and come back, take the new entry date as
first contract date instead of the original one.

closes odoo/enterprise#23790

Taskid: 2745947
X-original-commit: 8db4050ead0cd88b6cdef07af07e0cb833de8c06
Related: odoo/odoo#83493
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>