[IMP] web_dashboard: set min-height on .o_graph_canvas_container
After the revamp of the Graph View, it now uses the h-100 class which
overwrites the height of .o_graph_canvas_container and breaks the layout
in the Dashboard View.

To fix that and to keep consistency with the layout, this commit
replaces the height with a min-height.

community: https://github.com/odoo/odoo/pull/83132
task-2740030

closes odoo/enterprise#23633

Related: odoo/odoo#83132
Signed-off-by: Adrien Dieudonné (adr) <adr@odoo.com>