[FIX] delivery_ups : change default ups package dimensions
To reproduce:
Publish UPS US with the package type UPS Package/customer supplied
Try to send the package to a us address
Error "Error: Package exceeds the maximum length constraint of 108 inches. Length is the longest side of a package."

When dimensions are not provided, the default package is chosen, which is 520x520x520 mm.
The user can select between two dimension units (inches or centimeteres) in shippement method.
If the user chose inches (same thing for CM), the ups API is called with { dimensions : 520x520x520, unit : IN }.
Which exceeds the length limit 108 IN.

To fix that, the package 520x520x520 (MM) has been changed to 52x52x52 (CM).
So that even if the user chose IN as default unit that will not exceed the 108 IN limit

opw-2733844

closes odoo/enterprise#24088

X-original-commit: 68cbeb90ab122e60a509a4e6414b042c626bd8be
Signed-off-by: William Henrotin (whe) <whe@odoo.com>
Signed-off-by: abla001 <abla@odoo.com>