[FIX] account_reports: fix audit on general ledger
The general ledger is showing tax reports line when only selecting
a single journal of type sale or purchase.
The issue is that the method called only exists on the tax report and
not the general ledger, so it gives a traceback.
This is fixed by adding a method calling the correct one in the
tax report from the general ledger.

closes odoo/enterprise#23811

X-original-commit: e03e7a55f105d87a5e404b878f02b4e812354aed
Signed-off-by: Quentin De Paoli <qdp@odoo.com>
Signed-off-by: Nicolas Viseur <vin@odoo.com>