[IMP] delivery_dhl: input custom data
Before this commit, there was no way to insert custom data to the
DHL connector. With this commit, we add a custom field that can be
accesses from the delivery carrier view, only in debug mode.

This custom field will be evaluated like the inside of a python dict().

The keys of this dict will be 'rate', 'ship', 'return' depending on
when/where the user wants to add custom data.

Just before we send the data to the API, the request items are of type
`<class 'zeep.objects.Request'>`. This is practical to keep it that way
until then because the data added to it is checked with the wsdl file.

However, this is not practical when we want to add custom data. This
is more convenient to use a dictionary at this point. Dictionaries are
supported by zeep, so we do not need to convert it back to a zeep
object. to convert the zeep object to a dictionary, we can use the
function `zeep.helpers.serialize_object`. By default, the resulting
object is an ordered dict, which is an instance of a dict.

example of valid custom data:
```
'rate': {
    "Request": {
        "MetaData": {
            "SoftwareName": "DefinitelyNotOdoo",
        }
    },
},
'ship': {
    "ShipmentDetails": {"Pieces": {"Piece": {"Weight": '0.03'}}},
},
```

In the example for the 'ship' request, the Piece object is actually a
list of Pieve object. Each Piece object will be affected by the new
value.

task-2658889

closes odoo/enterprise#24224

Signed-off-by: Arnold Moyaux <arm@odoo.com>