[FIX] hr_contract_salary: Fix rounding issue on signature
Purpose
=======

In the US salaries are always discussed on a yearly basis. The "Generate offer" wizard supports this
but it only uses the yearly cost to calculate the monthly wage field. final_yearly_costs is then
calculated from the monthly wage field. This leads to rounding errors.

When I create an offer via the "Generate offer" wizard for $100.000 yearly the salary configurator
and contract end up like this:

- 8333.33 (for 8333.3333333)
- 99999.96 (for 100000)

To avoid confusion for the end users, even if this behavior is logical and not severe (4 cents dispatched
over a whole year), this commit prevents the amount to be modified no a nearby value.

Of courses, writing on an advantage will trigger the recompute to be set to the new value.

closes odoo/enterprise#23868

X-original-commit: a6f4b2e77d43b4853d43c56d9151bd167297e1ab
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>