[IMP] appointment: add a badge with number of slots of others months
When sharing a custom appointment link, we can have a situation in
the end of a month where it's hard to know at first sight if there
are other slots available.
We add a badge with the number of slots in the next and previous
months available.
This badge is only visible for custom appointment type.

task-2744256

closes odoo/enterprise#23704

Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>