[IMP] sale_{renting,subscription,temporal}: adapt to new price computation
This commit adapts sale temporal so that temporal product prices are
computed through the same methods than other products, requiring the
definition of a start_date and end_date to work as rental product, or a
unit and duration for subscription products.

This allows all the methods that uses product price methods on the
pricelist to access the right price without defining specific
inheritance on temporal products.

Defaultly, the price of a temporal product is the first price of its
product pricing for a given pricelist.

task-2382027