[IMP] (website_)appointment: improve calendar front end design
In master, the calendar UI to register for an appointment is unclear.

In this commit, we improve the design for visitors in a few ways:

 - When a month has no available slot, a descriptive message is shown
on the right side, with a relevant icon, and a link to the first
available slot.
 - When an appointment type has no available slot altogether, a message
is also displayed on the right side, with a relevant icon, and a link
to the 'contact us' page if website is installed.

We also improve the experience of logged-in users:

 - When an appointment type has no available slot, a relevant message
is shown on top of the calendar, specifying why there aren't any
(no user, no availabilities configured,...). A link to the back
end is also added.

task-2501575

closes odoo/enterprise#21548

Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>
Co-authored-by: Philémon van Helden <pvh@odoo.com>
Co-authored-by: Jérémy Hennecart (jeh) <jeh@odoo.com>