[IMP] l10n_cl_edi: reverse and debit note wizard should be country specific
We take the document type from the wizard instead of recalculating it again.

Before, the wizard calculated the document type by itself, because we thought or to
be only one for the reverse.  But a logic is already there in l10n_latam_invoice_document
 to support multiple and it calculates the correct value already.  So, we let that
module just do its job.  This way the user can now select another document type if needed.

Task: 2751825
X-original-commit: 9e049b833b23809e4287ef61ddf7482413debefd
Part-of: odoo/enterprise#25139