[FIX] documents_spreadsheet: preserve domain on queries to order values
When asking the server for ordered values on the model of the pivot, we
would ignore the original domain of the pivot, which could make some
queries explode (i.e. search every lead of type 'lead' or 'opportunity'
on a database with millions of records) and fail over a MemoryError.

This commit ensures that we preserve the original domain of the pivot.

task 2752189

closes odoo/enterprise#24450

X-original-commit: 2fb09129c7561a141b753767b9817df633e1139f
Signed-off-by: Pierre Rousseau (pro) <pro@odoo.com>
Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>
Co-authored-by: Lucas Lefèvre <lul@odoo.com>
Co-authored-by: Alexis lacroix <laa@odoo.com>