[IMP] delivery_fedex: add shipping insurance
Shipping insurance is a service which may reimburse senders whose parcels are
lost, stolen, and/or damaged in transit. With FedEx, we can declare a value
for each package. This value corresponds to the cost of the products in
the package.

Since the currency could vary from one picking to another, the currency
in which the declared values are sent is the currency of the company of
the delivery package.

FedEx has some limitations to the maximum insured value:
 - The maximum declared value is $1,000USD for FedEx First Overnight, FedEx Priority Overnight, FedEx
   Standard Overnight, and FedEx 2Day.
 - The maximum declared value is $50,000USD for FedEx 2Day Freight and FedEx 3Day Freight.
 - The maximum declared value is $100USD for FedEx Ground and FedEx Home Delivery.
These limitations will raise the UserError:
"8443: Package 1 - Insured Value exceeds limit allowed."

task-2701459
PR in base: https://github.com/odoo/odoo/pull/83896

closes odoo/enterprise#23708

Signed-off-by: Arnold Moyaux <arm@odoo.com>