[IMP] l10n_cl_edi: add locking mechanism to avoid concurrency issues
Before, it was possible that the sending of the invoice would be
triggered at the same time, often leading to the first process
putting the invoice to accepted, but the second one putting it back
to rejected, because the SII had already received it.

By locking the invoice, we prevent 2 processes from both sending
the invoice to the government.

Also avoid double send on double click of the send link to the SII, by
checking the l10n_cl_dte_status

opw-2674200
latam task 686 617

closes odoo/enterprise#24551

X-original-commit: 7df8115ad891f0771cedd9fce056b98425f443f2
Signed-off-by: William André (wan) <wan@odoo.com>