[IMP] payment_sepa_direct_debit: search only once for the transaction
Before this commit, most acquirers needed to run several successive
searches for the transaction whose reference was received by a
controller in notification data. This is because the security checks
run on the notification data require access to the acquirer through the
transaction record which was immediately discarded.

Starting with this commit, all `*_feedback_data` method are no longer
decorated with `api.model` and can use the transaction record they're
called on if provided. They are also renamed to `*_notification_data`.

task-2737144

closes odoo/enterprise#23938

Related: odoo/odoo#83850
Signed-off-by: Antoine Vandevenne (anv) <anv@odoo.com>