[FIX] industry_fsm_sale: update search pricing_type method for fsm
Before this commit, the `_search_pricing_type` method is not overriden
in `industry_fsm_sale` module like is made for the
`_compute_pricing_type` method.

This commit overrides the `_search_pricing_type` method to fit the
domain built for also the fsm projects and not only for the basic
projects.

The issue has been found during the development of task-2680425

closes odoo/enterprise#23951

X-original-commit: 2dd360206b66f950cfc4ee94049205bfd031b17f
Related: odoo/odoo#83874
Signed-off-by: Laurent Stukkens (ltu) <ltu@odoo.com>