[FIX] planning: fix the sample data visibility on switch view
Currently, the sample data is not displayed when switching views
in planning due to the invalid context in the reload method.
we already passing the show_job_title context in _fetchData
but when we reload it we are not passing the same context.

so in this commit, pass the show_job_title context in
__reload method to display the sample data properly.

task-2720976

Part-of: odoo/enterprise#23884