[FIX] l10n_pe_edi: Improve error message for OSE error 2022
We provide "helpful" error messages for some of the more typical error
codes returned by the OSE. Code 2022 is one of those, which triggers
when the partner name does not meet some validation criteria.

From testing, it seems that these validation criteria include:
- the partner name is at least 2 characters long
- the partner name does not contain special characters e.g. a tab.

At the moment, our "helpful" error message tells the user
to change the partner name to contain at least 2 characters. This
can be confusing if the problem with the partner name is in fact
that it contains a special character.

This commit amends the helpful error message so that it suggests
the possibility that the partner name contains a special character.

See this video where I show reproduction steps -> https://drive.google.com/file/d/1UVuqdo9__yV4Euk6XEn55oHmFAo8691I/view?usp=sharing

closes odoo/enterprise#29653

X-original-commit: 48e97c863f64c6d7a98f839bb6d6add64ce28f2e
Signed-off-by: Josse Colpaert <jco@odoo.com>
Signed-off-by: Antoine Dupuis (andu) <andu@odoo.com>