[FIX] website_calendar: fix double click on submit button
When we click multiple times on the submit button, an appointment
can be created a second time before the user can be redirected to the
confirmed paged. In this case, the appointment is created but the user
is redirected to the slot selection because it says that we already
have an event planned (the appointment which has just been created)
when the controller try to create a second one with the mutli-click.

To avoid this we use the automatic system making handlers promise-ready
(see makeButtonHandler() function for more info) and return a Promise
that is never resolved to prevent sending the form multiple times.
This Promise is sent only when the form is valid.

task-2750701

closes odoo/enterprise#25701

X-original-commit: 036e8368cfe6fba6e1856c355117e7fcda8f9bf7
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>