[IMP] utm: give same rights for utm campaigns to internal/social users
Currently social user have only read access for the utm campaigns,
however internal users read/write/create access for the same.

However, UTM campaigns are used a bit everywhere to it's okay to make
it a bit open. So this commit updates the rights of utm campaigns for
social users to make it same as internal users.

Task-id-2701582

closes odoo/enterprise#24124

Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>