[FIX] hr_payroll: avoid following contract when indexing
current behavior :
When a user indexes one or more contracts, they will automatically be set as a follower of those. Since a user usually indexes a lot of contracts at once, they will be spammed with notifications and will need to manually unfollow every contract.

behavior after this commit :
You are not added as a follower of contracts that you index.

taskID : 2729661

closes odoo/enterprise#24139

X-original-commit: 74d9496af6ab0ed92388c1de11502cd8a990d40b
Signed-off-by: Kevin Baptiste <kba@odoo.com>