[FIX] account_bank_statement_import_qif: Allow QIF files starting with a newline
Importing QIF file starting with a newline (\n) to create a bank statement will trigger an error
"Could not decipher the QIF file.".
This can cause issues because some banks (like "Chase Bank") start their file with a newline.

This fix will strip the file data before process it to avoid this.

opw-2739328

closes odoo/enterprise#24238

X-original-commit: b8407c4a7f2cb5ecf84e95baa6e14a0a3da81481
Signed-off-by: Cedric Snauwaert <csn@odoo.com>