[FIX] appointment: avoid redundant string in email
Before this commit,
While booking and cancelling the appointment, the email being set has some
duplicated content, for example, "Meeting with UserName with Username".
This happens with the custom invitation which already contains the
username (by default).

To fix this, either we can improve the default name of the custom invitation
or we can check in the template whether the invitation is custom. But with a
later approach, there could be an issue because the default name can be changed
by the user so we won't have proper wording in the email.

After this commit,
1) Change the default name of the custom types so that there is no "with"
   e.g: Username - Let's meet
2) Remove the "with Username" from the template for the "custom" types.

Task-2713963

closes odoo/enterprise#23197

Related: odoo/odoo#82894
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>