[FIX] account_invoice_extract: don't set customer invoice name
Setting the name of a customer invoice might change the sequence
associated to it.
This is very problematic as it's practically invisible to the user
and difficult to fix once it has been changed.

From now on, the name won't be set, the current sequence will be
used to generate it.

opw-2734614

closes odoo/enterprise#24092

X-original-commit: 13599b3eab3f4e21b45e8f10f34941e1b1ef100b
Signed-off-by: Florian Daloze (fda) <fda@odoo.com>
Signed-off-by: Louis Baudoux (lba) <lba@odoo.com>