[IMP] project_timesheet_forecast: sort dates in decreasing order in pivot view
In the project 'Timesheets and Planning Analysis' report, the hours
are grouped by date by default. Thoses dates are sorted in increasing
order, which can be a problem with a large amount of dates, since the
most relevant ones will be at the bottom.

To improve this, this commit sorts the dates by decreasing order instead.

Task-2741733

Part-of: odoo/enterprise#25080