[IMP] pos_loyalty: test to check merging of partner
The patch has been implemented in the odoo base code but the bug
has been discovered by using this module. That's why the test has
been added in this module.

How to reproduce the bug:

- Install POS and data_cleaning
- Create 2 customers and set loyalty points
- Select both customers and merge them
- Once the merge is done, the loyalty points have not been summed

Bug:

When you merge 2 contacts that have both loyalty points, the resulting
contact won't have the sum of the 2 source contacts. Instead of that,
the resulting contact will only have the points of one of of the
contact.

opw-2686208

closes odoo/enterprise#24445

X-original-commit: 5c7df0947922056a5c21c6934298bf9855d23614
Related: odoo/odoo#84739
Signed-off-by: Simon Goffin <sig@odoo.com>
Signed-off-by: Raphael Collet <rco@odoo.com>