[FIX] sale_renting : display correct return time in sol description
Steps :
Create a Rental Order.
Add a SOL with Product Meeting Room.
In the wizard, as Dates : Tomorrow 10:00 - Tomorrow 12:00.
Validate.

Issue :
See, in SOL Description : "Meeting Room 02/08/2022 10:00:00 to 13:00:00".
Expected : "... to 12:00:00".

Cause :
In previous versions, user tz was not taken into account in format_time,
so we used to first convert it to user tz before passing it to this method.

Fix :
Now, user tz is taken into account in format_time, so we don't convert
it.

closes odoo/enterprise#24305

Opw: 2746372
X-original-commit: b8bcc2a6ef08e932794c81876b0881682ea8daad
Signed-off-by: Onockx Audric (auon) <auon@odoo.com>
Signed-off-by: Victor Feyens (vfe) <vfe@odoo.com>