[IMP] account_invoice_extract: add separate settings for OUT invoices
The settings that determines the digitalization mode (disabled, manual
or automatic) used to be common for both vendor bills and customer
invoices.
As many users only want to use the service on vendor bills and not on
customer invoices, the settings has been split for both types of
invoices.

closes odoo/enterprise#24083

Task: 2751408
Related: odoo/odoo#84184
Related: odoo/upgrade#3240
Signed-off-by: Florian Daloze (fda) <fda@odoo.com>