[FIX] sale_renting : display one date when same day rental (tz)
Steps :
Create a Rental Order.
Add a SOL with Product Meeting Room.
In the wizard, as Dates : Tomorrow 00:00 - Tomorrow 23:00.
Validate.

Issue :
See, in SOL Description : "Meeting Room 02/08/2022 00:00:00 to 02/08/2022 23:00:00".
Expected : "...to 23:00:00".

Cause :
If picking date is equal to return date, the date should only be
displayed once. Yet, the dates are not timezoned for this check.

Fix :
Take timezone into account to check this condition.

closes odoo/enterprise#24162

Opw: 2746372
X-original-commit: 15785a2ea3b69a8e7cedf0f9657b423f06c323ce
Signed-off-by: Victor Feyens (vfe) <vfe@odoo.com>
Signed-off-by: Onockx Audric (auon) <auon@odoo.com>