[FIX] sale_timesheet_enterprise: get validated timesheets in project sharing
Before this commit, when the "Invoicing Policy" in Timesheet
app (that "sale.invoiced_timesheet" param) is equal to "approved",
only the validated timesheets should be visible for the portal user if
he has access to the related task. In project sharing, it is not the
case, the portal user sees no timesheets related to the task when the
`sale.invoiced_timesheets == 'approved'`.

This commit fixes the issue to show the validated timesheets related to
a task when the user is in the task form view via the project sharing
feature.

(Bug discovered during the development of task-2648955)

closes odoo/enterprise#24207

X-original-commit: 261d3c4916e6d5400df87706e378d21aa3069ba4
Related: odoo/odoo#84298
Signed-off-by: Laurent Stukkens (ltu) <ltu@odoo.com>