[FIX] helpdesk: allow overriding portal domain
Before this commit the only way to modify the domain is to completely override my_helpdesk_tickets.
Since this function is so big this is not clean/easy to do.
By creating a separate function we can simply override it and we can reuse the same domain in two places.

closes odoo/enterprise#25076

X-original-commit: 77569c3ceb700c2c00953b8a21226f7b324cabea
Signed-off-by: Laurent Stukkens (ltu) <ltu@odoo.com>