[FIX] l10n_nl_reports: XAF auditfile not exporting (timeout)
The goal is to improve the XAF export's performance in `l10n_nl_reports`.
Currently, it can lead to timeouts on large databases.

A test is added as well, to make sure the file generated
is identical before and after the refactor.

In `master`, the old template is removed so that the new one
can replace it from now on.

opw-2703534

X-original-commit: 8c0d73af165b0383ea57b69bbb46f7be3f641ddd
Part-of: odoo/enterprise#26531