[FIX] stock_account_enterprise: test ok if archived product
Change that archive/delete variant that are excluded will archive
a product ("[DESK0004] Customizable Desk (CONFIG) (Aluminium, Black)")
in demo data (odoo/odoo#83427).

With this change, stock_account_enterprise TestStockReport fails
because stock.report is compared to products and stock.valuation.layer
that are not archived.

With this changeset, the test is updated so it still works even if used
demo data are archived.

opw-2729329

closes odoo/enterprise#24512

X-original-commit: c87ede253b4928d87254003a91c3a2114c7a5ba7
Related: odoo/odoo#84883
Signed-off-by: Nicolas Lempereur (nle) <nle@odoo.com>
Signed-off-by: Sébastien Theys (seb) <seb@odoo.com>