[FIX] web_enterprise: Doesn't display studio button with invalid action
Currently when an action is invalid (for example if it does not contain a valid xml_id), the studio button in the systray is disabled to prevent the user from opening it.

https://github.com/odoo/enterprise/blob/5b111f2cfb1a577e75ccd8562adde4756dd8228c/web_studio/static/src/systray_item/systray_item.xml#L5

But he can still open studio via the "Add custom field" button in the list view.

And if it's an invalid view and the the user uses this button, a traceback `NotEditableActionError` is raised

https://github.com/odoo/enterprise/blob/b9d192067d289681c55948687ab6d3bc3a267d75/web_studio/static/src/studio_service.js#L139

This commit prevents the addition of this button in the optional fields dropdown in case the view is not editable by studio

opw-2686977

closes odoo/enterprise#24404

X-original-commit: f397e419de166ab4ab061dee44b31d75311aff0c
Signed-off-by: Achraf <abz@odoo.com>