[IMP] web_dashboard: update --PivotView-width
Prior this commit the width of the Pivot view was managed by different
width properties in css.
Now we use a css variable to manage the width depending on the context.
This commit updates this variable when the Pivot view is embedded in a
Dashboard view.

community: https://github.com/odoo/odoo/pull/82575
task-2731553

closes odoo/enterprise#23701

Related: odoo/odoo#82575
Signed-off-by: Adrien Dieudonné (adr) <adr@odoo.com>