[FIX] hr_referral: invalidate the cache after creating hr.referral.points
Purpose
=======
Now that the <mail.mail> are removed in a CRON, one test in hr_referral
failed. The reason is that the cache of `earned_points` need to be
invalidated after creating <hr.referral.points>.

Before, after the deletion of the <mail.mail>, the cache was removed
and that's why it worked.

Task-2587345

Part-of: odoo/enterprise#24266