[FIX] account_reports: fix carryover
Use case:
 Two lines as such:
  -27, using 22 in its formula and carrying over to 22 of the next period
   this carryover is not marked as persistent.
  -22 linked to a tax
Let's say, using an annual periodicity
In 2020, we put 100$ in 22. 27 balance would be 100. Post the closing entry.
In 2021, we find 100$ in 22 again (carried over) and the same amount in 27. Post the closing entry.
In 2022, the amount is now 0.

This happens because we check the difference between the current and past carryover,
which is 0 and then reset the carryover because it is 0.
But this doesn't make sense, because:
 - If the difference is 0, it means we have something to carryover
   just that it's the same amount as the last period, and so we do
   nothing.
 - If we don't carryover anything, the line amount would be 0, and thus
   the difference would already be -carryover_balance.

X-original-commit: b12557f6fb599b32afef8f37cc092190391a4609
Part-of: odoo/enterprise#24555