[IMP] hr_payroll: improve view of dashboard
- Change grid layout of dashboard in order to show notes in full-height.
- Changes text for warning view.
- Makes navigation between notes easier by showing the selected note like in
  other odoo notebooks (no border below the selected page).
- Allows deleting notes.
- Double-click now opens the configuration of a certain note. Single click
  just changes the active note.

task-2759601

[IMP] hr_payroll: improve note.note view

This commit turns the "add" button in the edit view for note.notes into a
secondary button (since there's already a primary button in the view) and
renames it to "share".

task-2759601

[IMP] hr_payroll: remove edit button in the dashboard notes

It is currently necessary to click on edit button to edit a certain note in
the dashboard. This is not very fluid and can be improved by entering edit
mode by clicking the wysiwyg editor. This commit implements this logic.

task-2759601