[FIX] documents_spreadsheet: missing nextTick in tests
Formulas `PIVOT`and `PIVOT.HEADER` have a potential async behaviour
since they require an RPC and a subsequent evaluation before they are
properly updated. Since ea20394, we introduced a debounce and with it
the necessity to run an additional nextTick to ensure that the cells
are evaluated after the RPC promise is solved.

In this commit, we use the helper `waitForEvaluation`that was added for
that very purpose.

task 2764761

closes odoo/enterprise#24348

X-original-commit: 4b01e34197feb7f1681e83b21b75dd3e5069a0bb
Signed-off-by: Pierre Rousseau (pro) <pro@odoo.com>