[REV] account_invoice_extract: don't set customer invoice name
This reverts commit 13599b3eab3f4e21b45e8f10f34941e1b1ef100b.

The commit is reverted as it wasn't the correct way to fix the issue.
The fact that the sequence was changed upon an OCR update shouldn't
be problematic as one shouldn't use the same journal to import
invoices from an external source and create them with Odoo. It should
be mutually exclusive.
The feature can be thus re-introduced.

closes odoo/enterprise#24411

X-original-commit: 6a976db9d7f23b6302300df8f1614236815cfba6
Signed-off-by: Florian Daloze (fda) <fda@odoo.com>
Signed-off-by: Louis Baudoux (lba) <lba@odoo.com>