[REF] documents_spreadsheet_bundle: Replace legacy domain code
Since Wowl, a new Domain class has been introduced which allows the
manipulation of domains expressed as strings or Arrays alike and
handles their combinations.

Switching to this new class makes the code more straightforward and easy
to read as we don't have to manually cast everything in order to use
pyUtils.

It also removes code that is now considered legacy.

Task 2762204

closes odoo/enterprise#24372

Signed-off-by: Pierre Rousseau (pro) <pro@odoo.com>