[IMP] helpdesk_sms, industry_fsm_sms: use sms template on task and ticket
The purpose of this commit is to include SMS templates. Templates allow people to
customize the content of sent SMS. When the task/ticket stage changes, the user will
be notified and there will be a track in the chatter.

- add this field on the right of the 'email template' one in list view.
optional and hidden by default.
- add it below the 'name' quick search in the search view.

Add SMS template data
  * Name: Field Service: Intervention Planned (see data)
  * Name: Field Service: Worker on its way (see data)

task-2683575

closes odoo/enterprise#22853

Related: odoo/odoo#81311
Signed-off-by: Laurent Stukkens (ltu) <ltu@odoo.com>