[FIX] web_enterprise: HomeMenu IME
Issue:

using special IME input does not work in several cases on home menu:

- pressing one key at page opening
- pressing one key after focusing on body (cliking on non-interactive
  element)
- pressing one key when an app or menu is focused

Example:

Pressing "a" key (on azerty, "q" on qwerty) on homepage while using
Traditional Cangjie IME will search "/手" in the command palette without
allowing the use of the IME.

=> In another input, the IME would open a graphical element allowing to
choose the symbol or to merge several symbols.

Cause:

An Input Method Editor only work on input fields (input, textarea,
element with contenteditable attribute), so we need to have focus on
such element or IME is disabled. So we lose focus when opening the
command palette.

Fix:

We have not found a solution that allows the command palette  to be opened
with the keypress while enabling the IME system. We decided that if the
keypress requires the use of the IME, the palette command will open
without the keypress value.

Note:

In the future there is a possibility that this could be solved with IME
API (https://www.w3.org/TR/ime-api/), but this is still in draft and
currently implemented in no modern browser (https://caniuse.com/ime).

With the current fix, if we focus on interactive element (eg. top-left
user menu) the IME is still disabled wrongfully.

This fix has been tested on Google Chrome, MacOS.

closes odoo/enterprise#24356

Related: odoo/odoo#84619
Signed-off-by: Géry Debongnie <ged@odoo.com>